move utf8decodebytes
