+++ lib/web/fetch/formdata-parser.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { webidl } = require('./webidl')
-const { utf8DecodeBytes } = require('./body')
+const { utf8DecodeBytes } = require('./util')
 const { collectASequenceOfCodePoints, removeChars, HTTP_TOKEN_CODEPOINTS } = require('./data-url')
 const { isFileLike, File: UndiciFile } = require('./file')
 const { makeEntry } = require('./formdata')
